Changed input pointcloud's is_dense flag to false
If the input pointcloud's is_dense flag is set to True, and it has invalid points (e.g. inf or NaN), the VoxelGrid filter incorrectly downsamples it. This addition ensures that no matter the is_dense flag, the pointcloud will always be treated as if it had invalid points, just in case.